Refactored and fixed bags

+++ KirovTransportTax.Domain/Entities/DriverTax.cs
@@ -3,6 +3,6 @@
     public class DriverTax
     {
         public string DriverPassport { get; set; }
-        public int SumTax { get; set; }
+        public float SumTax { get; set; }
     }
 }
